refactor(FNP-80): deleted authProvider from App, changed ProtectedRoute to use useGetUserQuery
