+++ packages/client/src/store/features/api.ts
@@ -5,4 +5,5 @@ import { API_URL } from 'src/utils/constants';
 export const api = createApi({
   baseQuery: fetchBaseQuery({ baseUrl: API_URL, credentials: 'include', mode: 'cors' }),
   endpoints: () => ({}),
+  tagTypes: ['Auth'],
 });
