+++ packages/client/src/store/index.ts
@@ -1,3 +1,2 @@
 export * from './store';
-export * from './api';
 export * from './slice';
